+++ include/rt_lighting.h
@@ -13,6 +13,6 @@
 #ifndef RT_LIGHTING_H
 # define RT_LIGHTING_H
 
-t_rt_color	precalculate_light(t_rt_obj_union *closest_obj, t_rt_vector o, t_rt_vector d, float closest_t, t_rt_scene *scene);
+t_rt_color	precalculate_light(t_rt_obj_union *closest_obj, t_rt_vector o, t_rt_vector d, double closest_t, t_rt_scene *scene);
 
 #endif
